airflow/requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-apache-hdfs
pyspark
